refactor(handlers): slim orchestrator below 600 lines
Sixth and final step of #529. Moves the orchestrator-owned handlers that
don't belong to any service (FormatMetrics, ExtractMarkdownSummary,
handleFileSelectionRequest, handleConversationSelected, handleMessageQueued,
HandleCommand) out of chat_handler.go into a sibling
chat_handler_handlers.go.

chat_handler.go is now 488 lines (down from the original 2,467). It
contains: the struct + constructor, the Handle() dispatcher, the
service-delegation wrappers, and the legacy-interface forwarders that
satisfy domain.ChatHandler.

Known follow-ups (out of scope for #529):
- ChatMessageProcessor and ChatShortcutHandler still hold *ChatHandler
  back-pointers. The refactor reduces what they reach for through it
  (only command-dispatch + a handful of injected services), but a clean
  fix requires replacing with explicit narrow deps. That is a
  self-contained follow-up.
- chat_shortcut_handler.go was 765 lines before this PR and remains so;
  splitting it by shortcut family (model, conversation, theme, a2a, etc.)
  is a separate concern.
